Created gui.py
THIS CODE IS VERY RAW.

All of the functionality is there, though it could use a bunch of testing. The code is horribly messy, and is in dire need of commenting and clean up. I'll get on that very soon! (Also I was unable to get the validation stuff to work, we should discuss that in our meeting, so I can fix it up).